Added precedence operator, modulus operattor, and improved error handling

+++ src/main.rs
@@ -13,20 +13,31 @@ use std::io;
 ///--------------------------------------------------------------------
 
 fn main() {
-    let mut translated_file = fs::OpenOptions::new()
+
+    let mut headers: bob::Headers = bob::Headers{ iostream: false, limits: false, string: false};
+    let arg: Vec<String> = env::args().collect();
+    let mut enable_source: bool = false;
+    // uses a reader buffer
+    if arg.len() < 2 {
+        panic!("Enter the name of the Bob source file!");
+    }
+    if arg.len() == 3 {
+        if arg[2] == "--output-source" {
+            if fs::metadata("output.cpp").is_ok() {
+                fs::remove_file("output.cpp").expect("Couldn't remove file output.cpp!");
+                println!("Bob replaced output.cpp!");
+            }
+            enable_source = true;
+        }
+    }
+        let mut translated_file = fs::OpenOptions::new()
         .read(true)
         .write(true)
         .create(true)
         .append(true)
         .open("output.cpp")
-        .expect("File creation failed");
-    let mut headers: bob::Headers = bob::Headers{ iostream: false, limits: false, string: false};
-    let arg: Vec<String> = env::args().collect();
-    // uses a reader buffer
-    if arg.len() != 2 {
-        panic!("Enter the name of the Bob source file!");
-    }
-    let reader = fs::File::open(&arg[1]).expect("Couldn't open that file");
+        .expect("File creation failed!");
+    let reader = fs::File::open(&arg[1]).expect("Couldn't open that file!");
     let reader = io::BufReader::new(reader);
     let mut variable_stack: Vec<bob::Variable> = Vec::new();
     for query in reader.lines() { // !! Is this the most efficient way to iterate through each line in the source? Also, read sentences instead of lines !!
@@ -36,7 +47,7 @@ fn main() {
             if query.len() == 0 {
                 continue;
             }
-            let query: String = lexical_analysis::string_space_remover(query);
+            let query: String = lexical_analysis::string_space_remover_and_bracket_replacer(query);
             let mut query: String = lexical_analysis::bob_and_punctuation_remover(query);
             let mut query_vector: Vec<String> = query.split_whitespace().map(String::from).collect::<Vec<String>>();
             bob::iterator(&mut query_vector,&mut translated_file, &mut headers,&mut variable_stack);
@@ -48,6 +59,8 @@ fn main() {
         .arg("-o")
         .arg("app")
         .status()
-        .expect("Couldn't run g++. Where's g++?");
+        .expect("Couldn't run g++. Where's g++ at?");
+    if !enable_source {
    fs::remove_file("output.cpp").expect("Bob couldn't delete his temporary file");
+    }
 }
